Move the new test case into qsimcirq_test.py
