--- qsimcirq_tests/qsim_circuit_test.py
@@ -1,41 +1,0 @@
-# Copyright 2019 Google LLC. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     https://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-import cirq
-
-import qsimcirq
-
-
-def test_qsim_circuit_eq():
-    q0 = cirq.LineQubit(0)
-    circuit = cirq.Circuit(cirq.H(q0))
-    qsim_circuit = qsimcirq.QSimCircuit(circuit)
-    qsim_circuit_2 = qsimcirq.QSimCircuit(circuit)
-
-    # Identical QSimCircuits are equal
-    assert qsim_circuit == qsim_circuit_2
-    assert qsim_circuit == qsim_circuit
-
-    # QSimCircuit and cirq.Circuit are not equal
-    assert qsim_circuit != circuit
-    assert circuit != qsim_circuit
-
-    # QSimCircuit and other types are not equal
-    assert qsim_circuit != None
-    assert qsim_circuit != "not a circuit"
-
-    # Different QSimCircuits are not equal
-    circuit_diff = cirq.Circuit(cirq.X(q0))
-    qsim_circuit_diff = qsimcirq.QSimCircuit(circuit_diff)
-    assert qsim_circuit != qsim_circuit_diff
